Added ellipsoid shape
Added an extra ellipsoid shape separate from the sphere. I have opted to
let the code that makes the sphere untouched via an if statement. If
all radii making up the ellipsoid are equal to each other, the sphere
is made via the old way. Otherwise an ellipsoid is created.